Lab 6-Answers
4.
a.FixedArrayQueue requires an explicit constructor because once the constructor
is defined in the class, the default constructor is not included. So, you have
to create the explicit construcotr in order to have it work the way the lab wants.

b. In my case, when testing this, the compiler gave me a warning. In a similar test,
it returns the number of elements in the queue. One example:

if(size==data.length) throw new IllegalStateException("Queue is full!);

So basically in this test, or in my specific situation here, it threw
the exception. It wont do this in every case obviously. If I were making
a queue prgram I would make sure to have a throw exception.

c. It returned a null in this case. The null is a special retrun value for the poll method
that tells you the queue is full. In most cases, you would have to have an isEmpty() method and have
it return null if the list for the queue is empty. However, the poll()
method can do that itself. It can see if the queue is empty, so it
depends on how you go about writing this nest of the code. I
had a return null; (needed for this, otherwise java will tell you
that you need to insert a return value).This can vary because Queue implementations
usually don't allow null elements to be inserted, but LinkedLists allow
you to do this.

d. It is constant which means it is of O(1) for the queue implementation and 0(n) for
the dequeue operation.