Question #1: Why does FixedArrayQueue require an explicit constructor? It is because FixedArrayQueue has a certain fixed size. There is a capacity in the previous lab that used linkedStacks.
Question # 2: What happens when you offer an item to a full FixedArrayQueue?
If you attempted to do this you would not be able to add another item since it is full. The Queue has reached its capacity, in this lab the capacity is 5.

Question #3: What happens when you poll an empty FixedArrayQueue?
It will throw java.util.NoSuchElementException. The explaination behind this is because the Queue is at zero.

Question #4: What is the time and (extra) space complexity of each of the FixedArrayQueue methods?
offer : O(1) as we are accessing the index in an array and array access is always O(1).
poll : O(1) same reason as offer
peek : O(1) same reason as poll
isEmpty : O(1) same reason as peek
size : O (1) same reason as isEmpty (all the same besides asList)
asList : O(n) as iterating through each element and adding that to the list.